Made minor improvmements
